--- src/lib/components/forms/index.ts
@@ -1,31 +1,0 @@
-// export { default as BloodOxygenSaturationForm } from "./response.types/advance/BloodOxygenSaturation.svelte";
-// export { default as BloodPressureForm } from "./response.types/advance/BloodPressure.svelte";
-// export { default as CholesterolForm } from "./response.types/advance/Cholesterol.svelte";
-// export { default as CReactiveProteinForm } from "./response.types/advance/CReactiveProtein.svelte";
-// export { default as ElectrolytesForm } from "./response.types/advance/Electrolytes.svelte";
-// export { default as GlucoseForm } from "./response.types/advance/Glucose.svelte";
-// export { default as HeightForm } from "./response.types/advance/Height.svelte";
-// export { default as HematocritForm } from "./response.types/advance/Hematocrit.svelte";
-// export { default as HemoglobinForm } from "./response.types/advance/Hemoglobin.svelte";
-// export { default as KidneyFunctionFaorm } from "./response.types/advance/KidneyFunction.svelte";
-// export { default as LipoproteinForm } from "./response.types/advance/Lipoprotein.svelte";
-// export { default as PulseRate } from "./response.types/advance/PulseRate.svelte";
-// export { default as RespiratoryRateForm } from "./response.types/advance/RespiratoryRate.svelte";
-// export { default as SleepForm } from "./response.types/advance/Sleep.svelte";
-// export { default as TemperatureForm } from "./response.types/advance/Temperature.svelte";
-// export { default as WaistForm } from "./response.types/advance/Waist.svelte";
-// export { default as WeightForm } from "./response.types/advance/Weight.svelte";
-
-// export { default as BooleanForm } from "./response.types/basic/Boolean.svelte";
-// export { default as CheckboxForm } from "./response.types/basic/CheckBox.svelte";
-// export { default as DateForm } from "./response.types/basic/Date.svelte";
-// export { default as DateTimeForm } from "./response.types/basic/DateTime.svelte";
-// export { default as FileForm } from "./response.types/basic/File.svelte";
-// export { default as FloatForm } from "./response.types/basic/Float.svelte";
-// export { default as IntegerForm } from "./response.types/basic/Integer.svelte";
-// export { default as ObjectForm } from "./response.types/basic/Object.svelte";
-// export { default as RadioForm } from "./response.types/basic/Radio.svelte";
-// export { default as RatingForm } from "./response.types/basic/Rating.svelte";
-// export { default as RangeForm } from "./response.types/basic/Range.svelte";
-// export { default as TextArrayForm } from "./response.types/basic/TextArray.svelte";
-// export { default as TextForm } from "./response.types/basic/Text.svelte";
