+++ src/lib/components/forms/responseType/DateFieldEditor.svelte
@@ -1,28 +1,21 @@
 <script lang="ts">
 	import { Input } from '$lib/components/ui/input';
 	import { Button } from '$lib/components/ui/button';
-	import { enhance } from '$app/forms';
 	import * as Card from '$lib/components/ui/card/index.js';
 	import { Label } from '$lib/components/ui/label';
 	import InfoIcon from '$lib/components/common/InfoIcon.svelte';
-	import { questionSchema } from '../question.schema';
-	// import InfoIcon from '';
-	// import { Label } from '../ui/label';
+	import type { QuestionUpdateModel } from '$lib/components/common/questionTypes';
+	import { questionSchema } from '../question-schema';
 
 	//////////////////////////////////////////////////////////////////////////////
 
-	let {
-		questionCard = $bindable(),
-		errors = $bindable(),
-		closeModel,
-		handleQuestionCardUpdate
-	} = $props();
+	let { questionCard = $bindable(), errors = $bindable(), handleQuestionCardUpdate } = $props();
 
 	async function handleSubmit(event) {
 		event.preventDefault();
 		console.log(questionCard.Title);
 
-		const model = {
+		const model: QuestionUpdateModel = {
 			id: questionCard.id,
 			Title: questionCard.Title,
 			Description: questionCard.Description,
@@ -30,9 +23,7 @@
 			Score: questionCard.Score,
 			CorrectAnswer: questionCard.CorrectAnswer,
 			Hint: questionCard.Hint,
-			QuestionImageUrl: questionCard.QuestionImageUrl,
-			RangeMin: questionCard.RangeMin,
-			RangeMax: questionCard.RangeMax,
+			QuestionImageUrl: questionCard.QuestionImageUrl ? questionCard.QuestionImageUrl : null,
 			IsRequired: questionCard.IsRequired
 		};
 
@@ -51,22 +42,8 @@
 			console.log('Called handleQuestionCardUpdate');
 			handleQuestionCardUpdate(model);
 		}
-
-		// const response = await fetch(`/api/server/question`, {
-		// 	method: 'PUT',
-		// 	body: JSON.stringify(model),
-		// 	headers: { 'content-type': 'application/json' }
-		// });
-		// const question = await response.json();
-		// console.log(question);
-		// if (question.HttpCode === 200) {
-		// 	closeModel('Card', question);
-		// }
 	}
 </script>
-
-<!-- method="POST"
-use:enhance -->
 
 <Card.Root class="rounded-lg border p-4">
 	<form
@@ -130,26 +107,6 @@ use:enhance -->
 			</div>
 		</div>
 		<Input bind:value={questionCard.ResponseType} class="hidden" />
-
-		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Minimum Range<span class="text-red-600">*</span></Label>
-			<div class="relative col-span-1">
-				<!-- Replace div with a button and handle keyboard accessibility -->
-				<InfoIcon title={'This is Question Score for Question.'} cls={'text-primary'} />
-			</div>
-		</div>
-		<Input bind:value={questionCard.RangeMin} type="number" />
-		<p class="error">{errors?.RangeMin}</p>
-
-		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Maximim Range<span class="text-red-600">*</span></Label>
-			<div class="relative col-span-1">
-				<!-- Replace div with a button and handle keyboard accessibility -->
-				<InfoIcon title={'This is Question Score for Question.'} cls={'text-primary'} />
-			</div>
-		</div>
-		<Input bind:value={questionCard.RangeMax} type="number" />
-		<p class="error">{errors?.RangeMax}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
 			<Label class="col-span-11 ">Score</Label>
