+++ src/lib/components/forms/responseType/ObjectFieldEditor.svelte
@@ -5,46 +5,54 @@
 	import * as Card from '$lib/components/ui/card/index.js';
 	import { Label } from '$lib/components/ui/label';
 	import InfoIcon from '$lib/components/common/InfoIcon.svelte';
-	// import InfoIcon from '';
-	// import { Label } from '../ui/label';
+	import type { QuestionUpdateModel } from '$lib/components/common/questionTypes';
+	import { questionSchema } from '../question-schema';
 
 	//////////////////////////////////////////////////////////////////////////////
 
-	let { questionCard, closeModel } = $props();
+	let { questionCard = $bindable(), errors = $bindable(), handleQuestionCardUpdate } = $props();
 
-	async function handleSubmit() {
+	async function handleSubmit(event) {
+		event.preventDefault();
 		console.log(questionCard.Title);
 
-		const model = {
+		const model: QuestionUpdateModel = {
 			id: questionCard.id,
-			title: questionCard.Title,
-			description: questionCard.Description,
-			responseType: questionCard.ResponseType,
-			score: questionCard.Score,
-			correctAnswer: questionCard.CorrectAnswer,
-			hint: questionCard.Hint,
-			questionImageUrl: questionCard.QuestionImageUrl,
+			Title: questionCard.Title,
+			Description: questionCard.Description,
+			ResponseType: questionCard.ResponseType,
+			Score: questionCard.Score,
+			CorrectAnswer: questionCard.CorrectAnswer,
+			Hint: questionCard.Hint,
+			QuestionImageUrl: questionCard.QuestionImageUrl,
 			IsRequired: questionCard.IsRequired
 		};
-		const response = await fetch(`/api/server/question`, {
-			method: 'PUT',
-			body: JSON.stringify(model),
-			headers: { 'content-type': 'application/json' }
-		});
-		const question = await response.json();
-		console.log(question);
-		if (question.HttpCode === 200) {
-			closeModel('Card', question);
+
+		const result = await questionSchema.safeParseAsync(model);
+		if (!result.success) {
+			console.log('client side validation error', result.error.flatten().fieldErrors);
+			errors = Object.fromEntries(
+				Object.entries(result.error.flatten().fieldErrors).map(([key, val]) => [
+					key,
+					val?.[0] || ''
+				])
+			);
+		}
+
+		if (Object.keys(errors).length === 0 || result?.success) {
+			console.log('Called handleQuestionCardUpdate');
+			handleQuestionCardUpdate(model);
 		}
 	}
 </script>
 
 <Card.Root class="rounded-lg border p-4">
 	<form
-		method="POST"
-		use:enhance
 		class="custom-scrollbar h-[calc(screen-2rem)] min-h-screen w-full overflow-y-hidden px-2 py-4"
-		onsubmit={handleSubmit}
+		onsubmit={(event) => {
+			event.preventDefault();
+			handleSubmit(event);
+		}}
 	>
 		<div class="relative mt-5 hidden grid-cols-12 items-center gap-4">
 			<Label class="col-span-11 ">Id</Label>
@@ -63,6 +71,7 @@
 			</div>
 		</div>
 		<Input bind:value={questionCard.Title} />
+		<p class="error">{errors?.Title}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
 			<Label class="col-span-11 ">Description</Label>
@@ -72,6 +81,7 @@
 			</div>
 		</div>
 		<Input bind:value={questionCard.Description} />
+		<p class="error">{errors?.Description}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
 			<div class="col-span-11 space-x-2">
@@ -100,40 +110,44 @@
 		<Input bind:value={questionCard.ResponseType} class="hidden" />
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Question Score</Label>
+			<Label class="col-span-11 ">Score</Label>
 			<div class="relative col-span-1">
 				<!-- Replace div with a button and handle keyboard accessibility -->
 				<InfoIcon title={'This is Question Score for Question.'} cls={'text-primary'} />
 			</div>
 		</div>
 		<Input bind:value={questionCard.Score} type="number" />
+		<p class="error">{errors?.Score}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Question Hint</Label>
+			<Label class="col-span-11 ">Hint</Label>
 			<div class="relative col-span-1">
 				<!-- Replace div with a button and handle keyboard accessibility -->
 				<InfoIcon title={'This is Question Hint for Question.'} cls={'text-primary'} />
 			</div>
 		</div>
 		<Input bind:value={questionCard.Hint} />
+		<p class="error">{errors?.Hint}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Question CorrectAnswer</Label>
+			<Label class="col-span-11 ">Correct Answer</Label>
 			<div class="relative col-span-1">
 				<!-- Replace div with a button and handle keyboard accessibility -->
 				<InfoIcon title={'This is CorrectAnswer for Question.'} cls={'text-primary'} />
 			</div>
 		</div>
-		<Input bind:value={questionCard.QuestionImageUrl} />
+		<Input bind:value={questionCard.CorrectAnswer} />
+		<p class="error">{errors?.CorrectAnswer}</p>
 
 		<div class="relative mt-5 grid grid-cols-12 items-center gap-4">
-			<Label class="col-span-11 ">Question QuestionImageUrl</Label>
+			<Label class="col-span-11 ">Question Image Url</Label>
 			<div class="relative col-span-1">
 				<!-- Replace div with a button and handle keyboard accessibility -->
 				<InfoIcon title={'This is Question QuestionImageUrl for Question.'} cls={'text-primary'} />
 			</div>
 		</div>
 		<Input bind:value={questionCard.QuestionImageUrl} />
+		<p class="error">{errors?.QuestionImageUrl}</p>
 
 		<Button type="submit" class="mx-auto mt-5 w-full">Add Question</Button>
 	</form>
